requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
nextcord==2.6.0
python-dotenv==1.0.0
sniffio==1.3.0
typing_extensions==4.8.0
yarl==1.9.2
